added MPS blur

+++ Athi Shared/Node/Quad.metal
@@ -59,14 +59,14 @@ fragment float4 gaussianBlurFrag(    Vertex              vert            [[stage
 }
 
 
-fragment half4 quadFrag(Vertex              vert           [[stage_in]],
-                        texture2d<half>     colorTexture   [[texture(0)]])
+fragment float4 quadFrag(Vertex              vert           [[stage_in]],
+                        texture2d<float>     colorTexture   [[texture(0)]])
 {
     constexpr sampler textureSampler (mag_filter::linear,
                                       min_filter::linear);
     
     // Sample the texture to obtain a color
-    const half4 colorSample = colorTexture.sample(textureSampler, vert.uv);
+    const float4 colorSample = colorTexture.sample(textureSampler, vert.uv);
     
     // We return the color of the texture
     return colorSample;
